Double 06-02-25

+++ Program.cs
@@ -11,9 +11,9 @@ namespace calculadoraSimples
         static void Main(string[] args)
         {
             Console.WriteLine("Digite o primeiro número");
-            int num1 = Convert.ToInt32(Console.ReadLine());
+            double num1 = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Digite o segundo número");
-            int num2 = Convert.ToInt32(Console.ReadLine());
+            double num2 = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine($"A soma do {num1} e {num2} é {num1 + num2}");
             Console.WriteLine($"A subitração do {num1} e {num2} é {num1 - num2}");
